refactor: adicionando @Builder nos models para facilitar criação de unit tests
