+++ src/main/java/br/com/roseai/sistemaloja/entity/Item.java
@@ -1,8 +1,7 @@
 package br.com.roseai.sistemaloja.entity;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.*;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
@@ -10,6 +9,9 @@ import java.time.LocalDateTime;
 
 @Getter
 @Setter
+@Builder
+@ToString
+@EqualsAndHashCode
 @Document
 public class Item {
