+++ src/main/java/br/com/roseai/sistemaloja/model/ItemDTO.java
@@ -2,11 +2,15 @@ package br.com.roseai.sistemaloja.model;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
+import lombok.ToString;
 import org.springframework.data.annotation.Id;
 
 @Getter
 @Builder
+@ToString
+@EqualsAndHashCode
 public class ItemDTO {
 
     @Id
